# feat(wmg): performance improvements