--- a/backend/wmg/api/common/rollup.py
+++ b/backend/wmg/api/common/rollup.py
@@ -10,7 +10,7 @@ import numpy as np
 import pandas as pd
 from pandas import DataFrame
 
-from backend.common.utils.rollup import rollup_across_cell_type_descendants
+from backend.common.utils.rollup import rollup_across_cell_type_descendants, rollup_across_cell_type_descendants_array
 
 ######################### PUBLIC FUNCTIONS IN ALPHABETIC ORDER ##################################
 
@@ -62,9 +62,7 @@ def rollup(gene_expression_df, cell_counts_grouped_df) -> Tuple[DataFrame, DataF
 ######################### PRIVATE FUNCTIONS IN ALPHABETIC ORDER ##################################
 
 
-def _add_missing_combinations_to_gene_expression_df_for_rollup(
-    gene_expression_df, universal_set_cell_counts_df
-) -> DataFrame:
+def _rollup_gene_expression(gene_expression_df, universal_set_cell_counts_df) -> DataFrame:
     """
     Augments the input gene expression dataframe to include
     (gene_ontology_term_id, tissue_ontology_term_id, cell_type_ontology_term_id, <compare_dimension>)
@@ -85,53 +83,58 @@ def _add_missing_combinations_to_gene_expression_df_for_rollup(
         and likely greater size than the input gene expression dataframe that includes combinations
         for which numeric values should be aggregated during the rollup operation.
     """
-    # extract group-by terms and queried genes from the input dataframes
-    # if a queried gene is not present in the input dot plot dataframe, we can safely
-    # ignore it as it need not be rolled up anyway.
-    group_by_terms = list(universal_set_cell_counts_df.index.names)
-    genes = list(set(gene_expression_df["gene_ontology_term_id"]))
+    if gene_expression_df.shape[0] == 0:
+        return gene_expression_df
 
-    # get the names of the numeric columns
     numeric_columns = list(
         gene_expression_df.columns[[np.issubdtype(dtype, np.number) for dtype in gene_expression_df.dtypes]]
     )
 
-    # exclude n_cells_tissue as we do not wish to roll it up
-    if "n_cells_tissue" in numeric_columns:
-        numeric_columns.remove("n_cells_tissue")
-
-    # get the total number of cells per tissue to populate the n_cells_tissue in the added entries
-    n_cells_tissue_dict = gene_expression_df.groupby("tissue_ontology_term_id").first()["n_cells_tissue"].to_dict()
-
-    # get the set of available combinations of group-by terms from the aggregated cell counts
+    group_by_terms = list(universal_set_cell_counts_df.index.names)
     available_combinations = set(universal_set_cell_counts_df.index.values)
 
-    # for each gene, get the set of available combinations of group-by terms from the input expression dataframe
-    entries_to_add = []
-    for gene in genes:
-        gene_expression_df_per_gene = gene_expression_df[gene_expression_df["gene_ontology_term_id"] == gene]
-        available_combinations_per_gene = set(zip(*gene_expression_df_per_gene[group_by_terms].values.T))
+    numeric_columns.remove("n_cells_tissue")
 
-        # get the combinations that are missing in the input expression dataframe
-        # these combinations have no data but can be rescued by the roll-up operation
-        missing_combinations = available_combinations.difference(available_combinations_per_gene)
-        for combo in missing_combinations:
-            entry = {dim: combo[i] for i, dim in enumerate(group_by_terms)}
+    pivoted = gene_expression_df.pivot_table(
+        index=group_by_terms, columns=["gene_ontology_term_id"], fill_value=0, values=numeric_columns, aggfunc="sum"
+    )
+    genes = np.array(pivoted["sum"].columns.tolist())
+    missing_combinations = list(available_combinations.difference(pivoted.index.values))
+    dense_tables = np.stack(
+        [
+            np.vstack((pivoted[col].values, np.zeros((len(missing_combinations), len(genes)))))
+            for col in numeric_columns
+        ],
+        axis=2,
+    )
 
-            # If a tissue, T1, DOES NOT have ANY of the QUERIED GENES expressed, then entirely
-            # omit all combos that contain tissue T1 from the candidate list of combos
-            # that should be considered for rollup. Otherwise, include combos containing T1
-            # in the candidate list of combos for rollup.
-            if entry["tissue_ontology_term_id"] in n_cells_tissue_dict:
-                entry.update({col: 0 for col in numeric_columns})
-                entry["n_cells_tissue"] = n_cells_tissue_dict[entry["tissue_ontology_term_id"]]
-                entry["gene_ontology_term_id"] = gene
-                entries_to_add.append(entry)
+    multi_index = pd.MultiIndex.from_tuples(pivoted.index.tolist() + missing_combinations, names=group_by_terms)
+    group_by_terms.remove("cell_type_ontology_term_id")
+    group_by_terms = ["cell_type_ontology_term_id"] + group_by_terms
 
-    # add the missing entries to the input expression dataframe
-    gene_expression_with_missing_combos_df = pd.concat((gene_expression_df, pd.DataFrame(entries_to_add)), axis=0)
+    multi_index = multi_index.reorder_levels(group_by_terms)
 
-    return gene_expression_with_missing_combos_df
+    multi_index_array = np.vstack(multi_index.values)
+
+    cell_types_for_rollup = np.char.add(multi_index_array[:, 0], ";;")
+    for col in range(1, multi_index_array.shape[1]):
+        if col > 1:
+            cell_types_for_rollup = np.char.add(cell_types_for_rollup, "--")
+        cell_types_for_rollup = np.char.add(cell_types_for_rollup, multi_index_array[:, col])
+
+    dense_tables = rollup_across_cell_type_descendants_array(dense_tables, cell_types_for_rollup)
+
+    row, col = dense_tables[:, :, 0].nonzero()
+
+    new_df = pd.DataFrame(index=multi_index[row], data=genes[col], columns=["gene_ontology_term_id"]).reset_index()
+    for i, col_name in enumerate(numeric_columns):
+        new_df[col_name] = dense_tables[row, col, i]
+
+    # get the total number of cells per tissue to populate the n_cells_tissue in the added entries
+    n_cells_tissue = gene_expression_df.groupby("tissue_ontology_term_id").first()["n_cells_tissue"]
+
+    new_df["n_cells_tissue"] = n_cells_tissue[new_df["tissue_ontology_term_id"]].values
+    return new_df[gene_expression_df.columns]
 
 
 def _build_cell_count_groups_universal_set(cell_counts_grouped_df) -> DataFrame:
@@ -208,11 +211,9 @@ def _rollup_cell_counts(cell_counts_grouped_df) -> DataFrame:
 
     if cell_counts_grouped_df.shape[0] > 0:
         universal_set_cell_counts_grouped_df = _build_cell_count_groups_universal_set(cell_counts_grouped_df)
+        index_names = universal_set_cell_counts_grouped_df.index.names
 
-        # Add index columns to the universal_set_cell_counts_grouped_df so that these columns can be
-        # DIRECTLY accessed in the dataframe during the rollup operation while traversing the cell type descendants
-        for col in universal_set_cell_counts_grouped_df.index.names:
-            universal_set_cell_counts_grouped_df[col] = universal_set_cell_counts_grouped_df.index.get_level_values(col)
+        universal_set_cell_counts_grouped_df = universal_set_cell_counts_grouped_df.reset_index()
 
         # rollup cell counts across cell type descendants
         rolled_up_cell_counts_grouped_df = rollup_across_cell_type_descendants(universal_set_cell_counts_grouped_df)
@@ -220,61 +221,6 @@ def _rollup_cell_counts(cell_counts_grouped_df) -> DataFrame:
         rolled_up_cell_counts_grouped_df = rolled_up_cell_counts_grouped_df[
             rolled_up_cell_counts_grouped_df["n_cells_cell_type"] > 0
         ]
-        # Remove columns that were added to the cell counts dataframe for the purpose of rollup.
-        # This is make it congruent with the structure of the input cell counts dataframe
-        rolled_up_cell_counts_grouped_df.drop(columns=rolled_up_cell_counts_grouped_df.index.names, inplace=True)
+        rolled_up_cell_counts_grouped_df.set_index(index_names, inplace=True)
 
     return rolled_up_cell_counts_grouped_df
-
-
-def _rollup_gene_expression(gene_expression_df, universal_set_cell_counts_df) -> DataFrame:
-    """
-    Roll up numeric values across cell type descendants in the input gene expression dataframe.
-
-    Accumulate gene expression values up the cell type ontology ancestor paths for every
-    (gene_ontology_term_id, tissue_ontology_term_id, cell_type_ontology_term_id, <compare_dimension>)
-    combination in the input gene expression dataframe. The compare dimension is an optional
-    field and in the case that it is missing, then accumulation happens for every
-    (gene_ontology_term_id, tissue_ontology_term_id, cell_type_ontology_term_id) combination.
-
-    For example:
-
-    If (G1, T1, C1) has gene expression values in the input expression dataframe,
-    and the tissue labeled T1 has another cell labeled C2 where C2 is an ANCESTOR of C1,
-    then the rolled up gene expression dataframe must include CUMULATIVE gene expression values for the
-    combination (G1, T1, C2) by adding in the gene expression for (G1, T1, C1).
-
-    Parameters
-    ----------
-    gene_expression_df : pandas DataFrame
-        Tidy gene expression dataframe containing the dimensions across which the numeric columns will be
-        aggregated.
-
-    universal_set_cell_counts_df : pandas DataFrame
-        Multi-indexed cell counts dataframe that contains "The Universal Set Of GroupBy Ontology Terms".
-
-    Returns
-    -------
-    rolled_up_gene_expression_df : pandas DataFrame
-        Tidy gene expression dataframe with the same columns as the input gene expression dataframe,
-        and likely greater size than the input gene expression dataframe, but with the numeric
-        columns aggregated across the cell type's descendants.
-    """
-    rolled_up_gene_expression_df = gene_expression_df
-
-    if gene_expression_df.shape[0] > 0:
-        # For each gene in the query, add missing combinations (tissue, cell type, compare dimension)
-        # to the expression dataframe
-        gene_expression_with_missing_combos_df = _add_missing_combinations_to_gene_expression_df_for_rollup(
-            gene_expression_df, universal_set_cell_counts_df
-        )
-
-        # Roll up expression dataframe
-        rolled_up_gene_expression_df = rollup_across_cell_type_descendants(
-            gene_expression_with_missing_combos_df, ignore_cols=["n_cells_tissue"]
-        )
-
-        # Filter out the entries that were added to the dataframe that remain zero after roll-up
-        rolled_up_gene_expression_df = rolled_up_gene_expression_df[rolled_up_gene_expression_df["sum"] > 0]
-
-    return rolled_up_gene_expression_df